requirements.txt
python-vlc==3.0.16120
python-socketio==5.7.2